refactor(backend): split media repository read and pipeline ops
